+++ frontend/src/types/booking.ts
@@ -1,0 +1,24 @@
+export interface CreateBookingRequest {
+  resourceId: string;
+  bookingDate: string;
+  startTime: string;
+  endTime: string;
+  purpose: string;
+  expectedAttendees: number;
+}
+
+export interface BookingResponse {
+  id: number;
+  resourceId: string;
+  bookingDate: string;
+  startTime: string;
+  endTime: string;
+  purpose: string;
+  expectedAttendees: number;
+  status: string;
+}
+
+export interface BookingFilters {
+  status?: string;
+  date?: string;
+}
